+++ CHANGELOG.md
@@ -4,6 +4,71 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.36.4-rc.0] - 2023-06-28
+### :bug: Bug Fixes
+- [`60d2017`](https://github.com/t3rn/t3rn/commit/60d2017fe52407ba1e0098ec8d51bb9ac74ee7d3) - update SDK metadata *(PR [#1095](https://github.com/t3rn/t3rn/pull/1095) by [@3h4x](https://github.com/3h4x))*
+
+### :wrench: Chores
+- [`5f76d56`](https://github.com/t3rn/t3rn/commit/5f76d56c4146bd1925190f544e13220d0c4c6cba) - **grandpa-ranger**: types/SDK changes should trigger deploy *(PR [#1099](https://github.com/t3rn/t3rn/pull/1099) by [@3h4x](https://github.com/3h4x))*
+
+
+## [v1.36.2-rc.0] - 2023-06-26
+### :bug: Bug Fixes
+- [`90bee60`](https://github.com/t3rn/t3rn/commit/90bee60e22a53cbc81c8daffeca9a897692611ad) - deploy phase should not be run by dependabot *(PR [#1085](https://github.com/t3rn/t3rn/pull/1085) by [@3h4x](https://github.com/3h4x))*
+- [`9d6b2e3`](https://github.com/t3rn/t3rn/commit/9d6b2e3ecb5a04005158ce4543fe911c78a2cf99) - executor deploy step *(PR [#1086](https://github.com/t3rn/t3rn/pull/1086) by [@3h4x](https://github.com/3h4x))*
+
+
+## [v1.36.0-rc.0] - 2023-06-21
+### :sparkles: New Features
+- [`e9e52ce`](https://github.com/t3rn/t3rn/commit/e9e52ce1b772da87348fe14b65e8a514ad9e732c) - add telemetry to executor *(PR [#1051](https://github.com/t3rn/t3rn/pull/1051) by [@ahkohd](https://github.com/ahkohd))*
+- [`d3672dd`](https://github.com/t3rn/t3rn/commit/d3672dd762cb1e7bcf7de958993e3d2caaad5b93) - process gateways activity overview in XDNS *(PR [#1054](https://github.com/t3rn/t3rn/pull/1054) by [@MaciejBaj](https://github.com/MaciejBaj))*
+
+### :bug: Bug Fixes
+- [`0f711aa`](https://github.com/t3rn/t3rn/commit/0f711aaeb5a5cc0779f5f8f5e9ab49bffa1e5895) - log all attester events and update metrics *(PR [#1078](https://github.com/t3rn/t3rn/pull/1078) by [@3h4x](https://github.com/3h4x))*
+
+
+## [v1.35.0-rc.0] - 2023-06-20
+### :sparkles: New Features
+- [`3ce9ff9`](https://github.com/t3rn/t3rn/commit/3ce9ff9c7e3aa05c712a7ba85b1e8631cd9095cb) - added attestationsDone array *(PR [#1075](https://github.com/t3rn/t3rn/pull/1075) by [@3h4x](https://github.com/3h4x))*
+
+### :wrench: Chores
+- [`e09d383`](https://github.com/t3rn/t3rn/commit/e09d383ed0e1646c48f0726e3e66687f07fdbf5a) - update CODEOWNERS and remove obsoleted files *(PR [#1076](https://github.com/t3rn/t3rn/pull/1076) by [@3h4x](https://github.com/3h4x))*
+- [`2466847`](https://github.com/t3rn/t3rn/commit/2466847b54e56c0eb9415630e07b91618cacb81d) - move pr template to t3rn repo *(PR [#1077](https://github.com/t3rn/t3rn/pull/1077) by [@3h4x](https://github.com/3h4x))*
+
+
+## [v1.34.2-rc.0] - 2023-06-19
+### :bug: Bug Fixes
+- [`85d9a26`](https://github.com/t3rn/t3rn/commit/85d9a26d3bd461348fdc0543285e26cf01b6127b) - remove unused labeled metric  *(PR [#1073](https://github.com/t3rn/t3rn/pull/1073) by [@petscheit](https://github.com/petscheit))*
+- [`42ef660`](https://github.com/t3rn/t3rn/commit/42ef6601fca29018b7116b422153e6a94efb374d) - quorum calculates Committee not ActiveSet size *(PR [#1074](https://github.com/t3rn/t3rn/pull/1074) by [@MaciejBaj](https://github.com/MaciejBaj))*
+
+### :wrench: Chores
+- [`eb92841`](https://github.com/t3rn/t3rn/commit/eb92841e2617badd58385e213423a8be0cb3312d) - run review only when label is added *(PR [#1072](https://github.com/t3rn/t3rn/pull/1072) by [@3h4x](https://github.com/3h4x))*
+
+
+## [v1.34.1-rc.0] - 2023-06-19
+### :bug: Bug Fixes
+- [`b79e72a`](https://github.com/t3rn/t3rn/commit/b79e72aeb89e67896a9d0b1510d30b5d66329d56) - skip decoding to H160 in attestation signature verification  *(PR [#1067](https://github.com/t3rn/t3rn/pull/1067) by [@MaciejBaj](https://github.com/MaciejBaj))*
+
+
+## [v1.34.0-rc.0] - 2023-06-19
+### :recycle: Refactors
+- [`293a7a6`](https://github.com/t3rn/t3rn/commit/293a7a6fed8cae763b62830eb653eeb38a579791) - customise max rewards executors kickback and default to zero *(PR [#1068](https://github.com/t3rn/t3rn/pull/1068) by [@MaciejBaj](https://github.com/MaciejBaj))*
+
+### :wrench: Chores
+- [`6b3b797`](https://github.com/t3rn/t3rn/commit/6b3b797dd4ee56fa2243f9650e899b7f0ce4a50d) - deploy docs on self hosted runners *(PR [#1071](https://github.com/t3rn/t3rn/pull/1071) by [@3h4x](https://github.com/3h4x))*
+
+
+## [v1.33.0-rc.0] - 2023-06-16
+### :sparkles: New Features
+- [`7e884bc`](https://github.com/t3rn/t3rn/commit/7e884bc7dc2218dad636a8c3a7fb9e003e560e91) - **attester**: submit attestation *(PR [#1037](https://github.com/t3rn/t3rn/pull/1037) by [@3h4x](https://github.com/3h4x))*
+
+### :bug: Bug Fixes
+- [`be15410`](https://github.com/t3rn/t3rn/commit/be154103fb85a2fdb3a522b4552f6542c96e5fa8) - sdk ignores local gateway entry *(PR [#1061](https://github.com/t3rn/t3rn/pull/1061) by [@petscheit](https://github.com/petscheit))*
+
+### :recycle: Refactors
+- [`37933ab`](https://github.com/t3rn/t3rn/commit/37933abb08ed6adb8c897c4f67c2a3d0ec838217) - check XDNS entry exists before adding as attestation targets + permanent slash as StorageValue *(PR [#1064](https://github.com/t3rn/t3rn/pull/1064) by [@MaciejBaj](https://github.com/MaciejBaj))*
+
+
 ## [v1.32.3-rc.0] - 2023-06-14
 ### :bug: Bug Fixes
 - [`b80f031`](https://github.com/t3rn/t3rn/commit/b80f0314fe46d68cea24287c8517258a4788aa55) - add reboot self gateway extrinsic to XDNS *(PR [#1060](https://github.com/t3rn/t3rn/pull/1060) by [@MaciejBaj](https://github.com/MaciejBaj))*
@@ -1232,3 +1297,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [v1.31.1-rc.0]: https://github.com/t3rn/t3rn/compare/v1.31.0-rc.0...v1.31.1-rc.0
 [v1.32.0-rc.0]: https://github.com/t3rn/t3rn/compare/v1.31.1-rc.0...v1.32.0-rc.0
 [v1.32.3-rc.0]: https://github.com/t3rn/t3rn/compare/v1.32.2-rc.0...v1.32.3-rc.0
+[v1.33.0-rc.0]: https://github.com/t3rn/t3rn/compare/v1.32.3-rc.0...v1.33.0-rc.0
+[v1.34.0-rc.0]: https://github.com/t3rn/t3rn/compare/v1.33.1-rc.0...v1.34.0-rc.0
+[v1.34.1-rc.0]: https://github.com/t3rn/t3rn/compare/v1.34.0-rc.0...v1.34.1-rc.0
+[v1.34.2-rc.0]: https://github.com/t3rn/t3rn/compare/v1.34.1-rc.0...v1.34.2-rc.0
+[v1.35.0-rc.0]: https://github.com/t3rn/t3rn/compare/v1.34.2-rc.0...v1.35.0-rc.0
+[v1.36.0-rc.0]: https://github.com/t3rn/t3rn/compare/v1.35.0-rc.0...v1.36.0-rc.0
+[v1.36.2-rc.0]: https://github.com/t3rn/t3rn/compare/v1.36.1-rc.0...v1.36.2-rc.0
+[v1.36.4-rc.0]: https://github.com/t3rn/t3rn/compare/v1.36.3-rc.0...v1.36.4-rc.0